refactor: 컴포넌트 폴더명 파스칼 케이스로 통일
