+++ src/components/Badges/FriendBadge.js
